Addon Preferences Created
Created Preferences folder, added addon_preferences to that

Added the development panel toggle to in out panel and the dev panel.

Added a comment to the dev panel
