+++ panels/inout_panel.py
@@ -10,6 +10,11 @@ class VIEW3D_PT_InOutPanel(bpy.types.Panel):
     bl_region_type = 'UI'
     bl_category = config["category"]
 
+    def poll(cls, context):
+        # Access addon preferences to check if the panel should be shown
+        prefs = bpy.context.preferences.addons[__name__].preferences
+        return prefs.enable_development_panel
+
     def draw(self, context):
         layout = self.layout
         props = context.scene.steps_props
